Move decorator and codeLensProvider to main directory.
